moving storage token verification to utils
