+++ server/routerlicious/packages/routerlicious-base/src/alfred/routes/api/documents.ts
@@ -4,30 +4,12 @@
  */
 
 import { IDocumentStorage, IThrottler, ITenantManager } from "@fluidframework/server-services-core";
-import { validateTokenClaims, throttle, IThrottleMiddlewareOptions } from "@fluidframework/server-services-utils";
-import { Request, Router } from "express";
+import { throttle, IThrottleMiddlewareOptions } from "@fluidframework/server-services-utils";
+import {  Router } from "express";
 import winston from "winston";
 import { IAlfredTenant } from "@fluidframework/server-services-client";
 import { Provider } from "nconf";
-import { getParam, Constants } from "../../../utils";
-
-// eslint-disable-next-line max-len
-async function verifyToken(request: Request, tenantManager: ITenantManager, maxTokenLifetimeSec: number, isTokenExpiryEnabled: boolean): Promise<void> {
-    const authorizationHeader = request.header("Authorization");
-    const regex = /Basic (.+)/;
-    const tokenMatch = regex.exec(authorizationHeader);
-    if (!tokenMatch || !tokenMatch[1]) {
-        return Promise.reject(new Error("Missing access token"));
-    }
-    const token = tokenMatch[1];
-    const tenantId = getParam(request.params, "tenantId");
-    const documentId = request.body.id;
-    const claims = validateTokenClaims(token, documentId, tenantId, maxTokenLifetimeSec, isTokenExpiryEnabled);
-    if (!claims) {
-        return Promise.reject(new Error("Invalid access token"));
-    }
-    return tenantManager.verifyToken(claims.tenantId, token);
-}
+import { getParam, Constants, verifyStorageToken } from "../../../utils";
 
 export function create(
     storage: IDocumentStorage,
@@ -45,7 +27,7 @@ export function create(
     router.get("/:tenantId?/:id", throttle(throttler, winston, commonThrottleOptions), (request, response, next) => {
         const maxTokenLifetimeSec = config.get("auth:maxTokenLifetimeSec") as number;
         const isTokenExpiryEnabled = config.get("auth:enableTokenExpiration") as boolean;
-        verifyToken(request, tenantManager, maxTokenLifetimeSec, isTokenExpiryEnabled).then(
+        verifyStorageToken(request, tenantManager, maxTokenLifetimeSec, isTokenExpiryEnabled).then(
             () => {
                 const documentP = storage.getDocument(
                 getParam(request.params, "tenantId") || appTenants[0].id,
@@ -71,7 +53,7 @@ export function create(
         const maxTokenLifetimeSec = config.get("auth:maxTokenLifetimeSec") as number;
         const isTokenExpiryEnabled = config.get("auth:enableTokenExpiration") as boolean;
 
-        verifyToken(request, tenantManager, maxTokenLifetimeSec, isTokenExpiryEnabled).then(
+        verifyStorageToken(request, tenantManager, maxTokenLifetimeSec, isTokenExpiryEnabled).then(
             () => {
                 // Tenant and document
                 const tenantId = getParam(request.params, "tenantId");
