+++ server/routerlicious/packages/routerlicious-base/src/utils/index.ts
@@ -5,3 +5,4 @@
 
 export * from "./constants";
 export * from "./params";
+export * from "./auth";
